feat: Add todo functionality to HomeView
This commit adds the functionality to add and remove todos in the HomeView component. It imports the HomeView component in the router and updates the template and script setup in HomeView.vue to include an input field to add new todos, a list to display existing todos, and a remove button for each todo item.